Write changelog to top of changelog when Previous Version is not found
As `previousVersion` would default to a non-falsey value if the provided
`--overwrite-version` did not exist, it would result in the new
changelog items being written to the _end_ of the changelog after
merging with the top/most recent changelog item and all previous
changelog items would be copied to the end of the file as well.

By defaulting to an empty string, we ensure that in cases where
`--overwrite-version` is a non-existing version (such as when being used
to set a specific, new version), we ensure that the changelog items are
able to be set at the top of the file and does not result in altering
any other entries.