Sort loaded types in Loaded Config
If a user-provided config defines its own, custom, types, they are
always placed at the end of the list shown via `npx
@stellarwp/changelogger add`.

This ensures that they are put in-place alphabetically with the default
types that are shown.

This does not impact how the changelog is written as that is handed via
the `WritingStrategy`.